sndFlags was called without initialised pointer
Fix regression introduced in 2003-01-08, with the addition of visible
tractors for observers.  The sndFlags function was being called with
a pointer that was not valid.  The results were unpredictable; on i386
the server state was corrupted, causing mass confusion of the clients,
and on sparc (from memory) cambot simply segfaulted.  Thanks to KeyOS
and the guys in IRC for ideas to debug.