describe a local server as on this computer
This is a refinement to the server multicast discovery that helps clients
to discover Netrek servers on a local area network.  One such scenario is
a server running on the same host as the client.  This server will now
answer with a comment to tell the user that the server is on the same
computer.