fix compilation errors on sequencer following previous commands.c change
The previous changes to commands.c to remove triple and balance votes
caused compilation errors on the sequencer because the functions had
gone away.  This showed that the functions are used by ntserv and the
sequencer, so they should move to libnetrek.

This patch moves the triple and balance voting code to a new file
balance.c in libnetrek.  It remains ifdef'd.