decide_ogg patch
The robots may know you carry. But once they decide to bomb, that's
all they will do.  They will cloak past an enemy carrier just to bomb.

Further, pops happen in groups, so there'd be 4 or more bot bombers.
All going for the same planet. 

This patch, makes it so, if a bomber's planet is already being bombed, 
instead of looking for another planet to bomb (which it can easily
find), it will instead look for close by carriers to ogg. 

It's not as extreme as it could be (switch to ogg mode right away, 
as soon as a carrier passes by right in front of you!!!). 
But, trust me, this patch makes the newbie's not want to play on a
newbie server.

Bombing and ogging should go hand-in-hand, they are similar types of 
work.