add setteam utility
Add a utility for viewing and changing the starbase reconstruction timer
and surrender timer, both of which are in struct teams.  Although these
timers can be manipulated by other means, setteam makes it easier to 
script.

Some common code now exists in setplanet.c and setteam.c that deserves
factorisation.