add install-ntserv target for live updates
Addition of a target install-ntserv to the ntserv Makefile so that a
server operator can do a live update of ntserv during a running game.

The target moves the existing in-use ntserv binary into an archive
tree before installing the new ntserv binary.  This makes it possible
to manually back out of an installation if the new binary breaks
badly.

Other methods to test a new ntserv binary include running it manually
for connection back to the client.  See ntserv command line options.