robotd_udcounter.dpatch
I re-did the setflag() function, so it now will always report back
a positive incrementing int at 100ms intervals. And it starts with 0
on program start. 

There are some other places that initialize _udcounter, I think this
might need to be got rid of, but I didn't do that in this patch. 

Read the longwinded diatribe in the Changelog.