fix SP_2 flag sending for real
The sndFlags function will not send flags of observers (except for
PFOBSERV of course).  The SP_2 flag sampling code will sample the flags
of robots now.  It wll not sample observers' flags, except for an
observer's own flags.  The SP_2 sampling code will update the last sent
flags data, so shield/cloak are not sent again via sndFlags().  When
sndFlags() does send flags, it will sent the correct shield/cloak and not
zero.