adapt when assaulting
The assault planet code didn't re-check the target planet after the
assault started.  If the planet was taken by another player on a friendly
team, or modified by god, the robot would still try to bomb/take it, even
when that action was no longer possible.  The robot will now check for
planets is can not drop on, and abort the assault if so.  It will not try
to bomb planets it can not bomb.  It will also stop reinforcing planets
when they get to 4 armies.