fix neutral planet check
Fix several bugs with check_take().  It didn't handle the case when no
enemy planets were left.  It looked at the last enemy planet in the list,
rather than the target enemy planet when deciding if it should take neut
planets.  It would skip looking for neuts when it found an enemy agri and
_didn't_ have enough armies to take it.