fix res danger
The check for res danger would make the robot cloak when near _any_ home
planet, so it would cloak even when bombing it's own core.  Unfortunately,
it is not possible to tell from the planet flags whose home world a
planet is.  Changed the code to only check the normal home planet
indexes, skipping the bots's own HW, for proximity.  This makes the code
much smaller and faster.  Realistically, the HWs are always at the
standard indexes, even with non-standard planet layouts.  Included a bit
of code, #if'ed out, that will find home worlds by name, the way the bot
does when it wants to refit.