torp dir for robot
Hadley's bot expects to get torp direction from the server, but the
server only sends it for your own torps.  Changed to send it for other
players' torps too.  Without it, the bot thinks all torps are going
straight up, making it a very bad dodger (unless you attack from below). 
  
This should be fixed to only turn this on for the robot, so as not to
help borgs.  This will have to wait on a better way to detect robots.