help client with plasma war in ping-pong mode
In Ping-Pong plasma mode, a plasma changes teams when it is bounced. 
There is no way to send this change to the client, so the client doesn't
know.  For example, if a player at peace with us but who we are at war
with bounces our (or a teammate's) plasma back at us, it appears to the
client that the plasma is friendly.  But really, it's not, since we are
at war with the plasma's new team.  Fix this by setting the plasma's war
flag if the player is hostile to the plasma's team.
 
In order to take advantage of this, clients will need to be sure to:
1. Not assume that the player's own plasma is friendly.
2. Not assume that plasma from the player's team is friendly.
3. Make sure the check the plasma's war flags, and not the flags of the
player who fired it.