help bot deal with ping-pong plasma
The robot has code to deal with Ping-Pong plasma, but it's an old
version.  It doesn't work properly with the current ping pong plasmas,
written by Trent Piepho in 1995 (so the old ones are really old!). 
Basically, the bots needs to allow for plasma fired by a friendly player
becoming hostile to it.
      
Also sets the pl_team member of the plasma struct, which makes getting
the plasma's team easier.