fix mp.h syntax error
Wasn't able to find what patch had caused this problem, darcs trackdown found
no patch removal that would pass this test, so it seems likely that it was due
to a change to autoconf in Debian Etch rather than in the Netrek Server Vanilla
code base.  Expanding the check into separate lines solved the problem.