netrekd, rewrite access checks to avoid DNS
This is a rewrite of netrekd access checks to avoid DNS.  Advantage is
much more rapid login, and honest reporting of source IP instead of
trusting the reverse DNS.  Disadvantage is revealing source IPs to all
players, and removing the sometimes useful hostnames from the login
or "@" message response.

The perceived regression may be fixed in a later patch.