blog, add minimal support for inl link to stats
Add an example of minimal support for INL blog, we need the server hostname
in a single place before it is practical to make this work well.  Note that
a make install is needed to create the blog/inl directory, but that the
directory can be removed if not required.