daemon, add SIGHUP handler to restart during play
Allows a server administrator to send a SIGHUP to the daemon during play
causing it to execute itself again.  Can be used to perform live code
changes to the daemon, provided the structure of the shared memory is
not changed.

Testing show far has shown no noticeable lag using a script that restarts the
daemon every second.  There may, however, be regressions, and so this patch
should be tested carefully.

Details of the change:

	* struct.h: move context below status so that a copy of status can
	be maintained in context, add daemon global variables to context,
	move player and torp arrays below status and context in memory
	layout to allow for future restart with memory layout changes per
	PROJECTS.checkpointing.

	* daemonII.c: move global variables into context struct, including
	tick cycle counter, remove some old code for rotating galaxy,
	define variables and new command line options, add a SIGHUP
	handler for restarting, do a restart immediately after a game move
	if SIGHUP is received, simplify removal of ONCHECK file using
	unlink instead of exec or fork, only send SIGUSR1 to parent if
	parent sends a --tell option to daemon when started.

	* openmem.c: when starting daemon, be sure to give --tell option.