Sturgeon refit and upgrade, the core of sturgeon
This patch contains the essential core of sturgeon.  With previous patches
establishing all the backend code to deal with changes to the player struct,
and backend for adding upgrades and such, this patch is the user interface
that all sturgeon players must navigate to add and remove upgrades.  Upgrading
is done through a series of menus and submenus that are activated by refitting
into a ship of the same class.  Starbases do not receive upgrades.  The menus
are navigated by the speed keys.  Herein lies the biggest design flaw of 
sturgeon, in that in order to make sturgeon comptable with clients, keys had
to be overloaded.  In the case of the coup key functioning as special weapons
toggle, or the plasma key firing all the different types of special weapons,
it is not that big a deal.  But for navigating the menus, the speed keys are
used.  Many older clients or players used to playing over a modem use
enforced UDP, which sends every speed packet *twice*.  This makes it
impossible to navigate the menus correctly.  Luckily, the number of players
affected by this (from my firsthand experience of running and providing
tech support for the sturgeon server) is small, and they are all older players.
New clients do not use enforced UDP by default.  Regardless, anyone running
a sturgeon server should at the very least update their MOTD with information
regarding enforced UDP problems.  See http://netrek.warped.us for the complete
sturgeon guide, including a section on troubleshooting.

   * interface.c (set_speed): Adds upgrade menus.  A typical upgrade involves
     navigating through 2 menus (i.e. selecting 2 speeds), then a check is
     made to see if the upgrade is valid.  In the case of the special weapons,
     a few other checks are made, such as to see if the ship already has an
     infinite amount of that weapon, or if it has the cargo space, in the
     case of nukes.
  *  (do_refit): Upon refitting to the same ship type (non-base) the first
     main upgrade menu is shown.  Included is an option for a normal refit.
     Upon successful refit, special weapons are reset.  Upgrades from the
     old ship are then applied to the new ship.  Before the undo upgrade
     option, all upgrades were refunded - that code is left in, but 
     commented out as undo upgrade proved more popular and a superior
     way of fine tuning a ship's upgrades.  The default ship special
     weapons are then applied.  Assault ships get unlimited mines, bases
     get unlimited pseudoplasma, type 5 plasma, and drones.  ATT gets
     unlimited everything, and GA gets unlimited pseudoplasma.  A ship
     cap packet is then sent.  Also note that plasma is handled differently
     in sturgeon - refitting with 2 kills does not grant plasma, all
     plasma must be bought, and it comes in limited amounts.
M ./Vanilla/ntserv/interface.c -25 +423