Sturgeon enter ship routine, the other half of sturgeon core
   * enter.c (enter):
This patch deals with what happens when a player dies.  It bears some
similarity to when a player refits.  Bases do not get to keep upgrades.
Depending on a few conditions, upgrades are lost.  Upon genocide or
conquer, upgrades are not lost.  If the player's rank credit is high
enough to cover all his upgrade costs, upgrades are not lost.  Otherwise,
upgrades are lost.  If sturgeon lite is set, only the costliest upgrade
is lost.  Otherwise, all upgrades are lost, but only down to the point
where rank credit is enough to pay for the remaining upgrades.  Thus,
in all cases, rank upgrades are essentially free upgrades that can't
be lost.  Rank credit is recalculated, and special weapons are reset,
then set to the specific ship values, as in refit().  Plasma is again
handled slightly differently in the sturgeon case, as sturgeon does
not use the plasmacost variable.

M ./Vanilla/ntserv/enter.c -3 +104