Sturgeon special weapon + plasmas, the messiest part of sturgeon code
* plasma.c (nplasmatorp);
This patch is the hardest part of sturgeon code to understand, as it
creates a complex tangle of nests within the main plasma function, which
is unavoidable as all special weapons are overloaded onto the plasma key.
This is also the likeliest place any bugs in sturgeon still exist.  Mines
and drones are assigned to the torp array, while plasmas are assigned
to the plasma array.  There is a known bug where a player can sometimes
fire an extra torp, in a different team color than his own.  I believe
there's some bug with the arrays, either here or in torp.c.  The checks
to see if a nuke can be used are here, including a large if/else  statement
which gives randomness to nuke events.
M ./Vanilla/ntserv/plasma.c -56 +313