pret, prevent kathy from exiting
It had been observed that kathy would quit out with no humans in the
game for 60 seconds and would not respawn until the daemon died and
restarted; this patch addresses that issue.  player bots will still exit
out as they should.  toggling of PRET=1 to PRET=0 in sysdef still causes
player bots and kathy to exit cleanly.