allow whitelist to override ignores
Allow whitelisted IP's to override player :ita settings.  Rational is that
whitelist entries are supposed to have a higher level of trust granted by
the server administrator and as such should be allowed to send messages to
players that have them :ita'd.  This patch allows for adminsitrator set
override levels for each of the 3 message classifications with all levels
defaulting to non-permissable override status.