remove lockout()
Removed "lockout()" checks from getname.c; if players are to be banned by
login it should not be a static check in the code but rather a dynamic 
configuration setup.  The code checked for "bozo" and locked out anyone
using that as a login string; I have been able to find to reason for this.