lame refit
Historically ships have been allowed to refit either on their homeworld or
while docked to their base with the following criteria: shields >= 75%,
fuel >= 75%, damage <= 75%.  I do not believe that this was the original
authors intent, as damage is the only of the three critera that increases
on a scale of good to bad; I believe the original intent was damage <= 25%.

It is the status quo for non-base ships to be able to refit while docked to
a base in clue games when heavily damaged.  But to allow a base to do so
while on home world is disruptive to the game.  The base is the most 
expensive of all ship classes and the only one with a 'regeneration'
period between base death and the availability of the next base. 

The current refit criteria is currently being abused on pickled to allow
core bases to refit on homeworld with significant amounts of damage, disrupting
the game and making bases virtually indestructable as a result of knowing
how to split damage to exploit this 'feature' of the code.

This patch has 2 configuration options to give server admins the ability
of 'fixing' the refit criteria and limit refits to those with <= 25%
damage individually for both bases (LAME_BASE_REFIT=0/1) and non-bases
(LAME_REFIT=0/1).  The patch defaults to the classical criteria to maintain
the status quo.