add ignoring/ignoredby commands
Added support for IGNORING and IGNOREDBY commands.  IGNORING will present a 
list of ip addresses the caller is ignoring, along with the slot of the ip if
it is currently in the game and the type of ignore in place.  IGNOREDBY
does the same for ip's that are ignoring the caller.