is_idle, not if torps in flight
Changes the definition of safe-idle to exclude ships that have torps in
flight, as it was possible to throw torps from the home world and yet be
safe from attack.

Adds GU_UNSAFE, which allows the client to see whether the ship is safe
idle, and as the bit used to be GU_GAMEOK the meaning is inverted.