decimal precision conformance for displayed stats
This patch addresses the disparity between decimal precision values for the player stats (Offense, Defense, Bombing, Planet) displayed in the playerlist and inform windows in clients versus those shown in the bounce session stats output the server sends a player when they send '?' for a player slot. The prior method displayed the stats using printf field width and precision modifiers which sometimes led to rounding errors. Precision is now maintained to two decimal places and the various ways these stats are displayed to the user are now uniform in the way the values are displayed. The new method uses the ftoa() function to generate values less susceptible to rounding errors.