add client name to version packet and log
Client normally identifies to server using a CP_MESSAGE on first
SP_PICKOK, but with a version number only.  This was used with RSA on
clients that did not support feature packets, and had no other use.

This change prints the whole version message to the ERRORS log, for
statistical analysis later.

Clients are encouraged to send "@name version", any number of spaces are
permitted, but the last word will be taken as the version number of the
client.

Also fixed a valgrind detected failure to free a heap pointer.