experimental contextual tips
Replaces the MOTD with a player performance tip specific to the context
in which the player died.  A list of several tips are triggered, with
the most important tip being sent.

May require client debugging to fix ClearMotd which has not previously
been used by the server.  If ClearMotd certainly cannot be made to work,
then this contextual tip feature may require feature packet support.