df-friendly galaxy generation
This adds a sysdef option to generate a dogfight-friendly galaxy: all
planets are created as INDI, 0 armies, REPAIR and FUEL and revealed to
all teams.  The aim is to generate a galaxy where players are able to 
fuel / repair anywhere, no matter what their war declarations may be
set as and precluding the chance of army dropping.   Also modified the
setgalaxy utility to provide a df-friendly reset option ("d").