Add additional offense restrictions
This patch adds offense restrictions for assault ships, battleships and 
scouts similar to the functionality that is already present for bases and
destroyers.  Defaults to legacy operation with no requirements in place.

This patch also endeavors to clean up some of the indenting in getentry.c
as what was present was quite out of sync.