teacher torps
An operating mode for use by a teacher in a classroom LAN game setting,
which either:

1.  denies the use of torpedos entirely, or;

2.  denies torpedo firing in green alert, or when the player has less
than one kill.

The intention is to remove the distraction of torpedos at the outset of
the training, and introduce them gradually.

The lesson plan would be something like (a) five minutes with PFWEP, (b)
five minutes with no torpedos, (c) five minutes with torpedos after
first kill, (d) unrestricted play.