Resolve all "implicit int" warnings raised by GCC.
These are primarily in the robotd code. They include such things as implicit int
as a return value, implicit int as a function parameter type, and implicit int
as a register type. The implicit int situation leads to cases where it isn't
clear whether a function is expected to return a value or not, which prevents
many of the compile time checks.