Fix global 'hours' buffer overflow
When loading the hours file, ntserv uses fscanf() to read play hours
line-by-line into a global 'hours' buffer. Since there are 24 hours in a
day, each day is an array of 24 hours. However, fscanf() also copies the
terminating NUL character. This results in a buffer overflow in the
default installation when parsing the last line.

This fix uses fgets() to read up to 25 characters (including the
trailing NUL), and then copies the first 24 characters into the global
'hours' buffer.