Solicit - do not count comment lines, log metaservers configured
The metaservers example file contains a number of explanative comments. Unfortunately the comments themselves were counted as entries for the metaserver array. Since there's a hard maximum of 8 servers, appending actual metaserver names at the end of the file results in ntserv solicit quietly failing. I've changed the code to simply not count comment lines as one of the 8 entries, which makes the default file work as intended. I did not want to introduce a lot of new changes, but I added some error log output that should be useful for startup.