Enable _FORTIFY_SOURCE on gcc and fix a buffer overflow
A buffer overflow does occur immediately after login in Tedd Hadley and
Heiko Wengler's amazing short packets code from 24th May 1993, which
relies on strcpy to overflow the destination string without complaint:

	strcpy(&dst->mesg, &src->m_data[9]);

The destination string is a one character buffer in the short packet
structure:

	struct mesg_s_spacket {
	   ...
	   char mesg, pad2, pad3, pad[76];
	}

Original intent was for the data to be written over the bytes in the
packet, overflowing into the remainder of the structure fields.

It is an intentional overflow, which the fortify wrappers are
reporting as a problem.

Change to use a char array.